Split EmojisLIstView and EmojisListViewController
