+++ Challange-1-iOS/Challange-1-iOS/Presentation/EmojisList/EmojisListView.swift
@@ -1,0 +1,73 @@
+//
+//  EmojiView.swift
+//  Challange-1-iOS
+//
+//  Created by Daniel Silva on 11/4/22.
+//
+
+import UIKit
+
+class EmojisListView: BaseGenericView {
+    var collectionView: UICollectionView
+
+    required init() {
+        let layout = UICollectionViewFlowLayout()
+        layout.scrollDirection = .vertical
+
+        layout.minimumLineSpacing = 8
+        layout.minimumInteritemSpacing = 4
+
+        collectionView = .init(frame: .zero, collectionViewLayout: layout)
+
+        super.init()
+        createSubviews()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    func createSubviews() {
+        setUpViews()
+        addViewsToSuperview()
+        setUpConstraints()
+    }
+
+    private func setUpViews() {
+        collectionView.register(EmojiCell.self, forCellWithReuseIdentifier: EmojiCell.reuseCellIdentifier)
+
+        collectionView.delegate = self
+    }
+
+    private func addViewsToSuperview() {
+        addSubview(collectionView)
+    }
+
+    private func setUpConstraints() {
+        collectionView.translatesAutoresizingMaskIntoConstraints = false
+
+        NSLayoutConstraint.activate([
+            collectionView.topAnchor.constraint(equalTo: topAnchor),
+            collectionView.bottomAnchor.constraint(equalTo: bottomAnchor),
+            collectionView.leftAnchor.constraint(equalTo: leftAnchor),
+            collectionView.rightAnchor.constraint(equalTo: rightAnchor)
+        ])
+    }
+}
+
+extension EmojisListView: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        insetForSectionAt section: Int) -> UIEdgeInsets {
+        return UIEdgeInsets(top: 1.0, left: 8.0, bottom: 1.0, right: 8.0)
+    }
+
+    func collectionView(_ collectionView: UICollectionView,
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        sizeForItemAt indexPath: IndexPath) -> CGSize {
+
+        guard let layout = collectionViewLayout as? UICollectionViewFlowLayout else { return .zero }
+        let widthPerItem = collectionView.frame.width / 5 - layout.minimumInteritemSpacing
+        return CGSize(width: widthPerItem - 8, height: widthPerItem)
+    }
+}
