+++ Challange-1-iOS/Challange-1-iOS/Presentation/EmojisList/EmojisListViewController.swift
@@ -1,0 +1,68 @@
+//
+//  EmojisList.swift
+//  Challange-1-iOS
+//
+//  Created by Daniel Silva on 9/26/22.
+//
+
+import UIKit
+
+class EmojisListViewController: BaseGenericViewController<EmojisListView>, Coordinating {
+    var coordinator: Coordinator?
+    var emojisList: [Emoji] = []
+    var viewModel: EmojiViewModel?
+
+    init() {
+
+        super.init(nibName: nil, bundle: nil)
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Emojis List"
+        genericView.collectionView.dataSource = self
+    }
+
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        viewModel?.emojisList.bind(listener: { [weak self] emojiArray in
+            guard let emojiArray = emojiArray else { return }
+            self?.emojisList = emojiArray
+            DispatchQueue.main.async { [weak self] in
+                self?.genericView.collectionView.reloadData()
+            }
+        })
+        viewModel?.getEmojis()
+    }
+}
+
+extension EmojisListViewController: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+
+        let countEmojis = emojisList.count
+
+        return countEmojis
+    }
+
+    func collectionView(_ collectionView: UICollectionView,
+                        cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+
+        let cell: EmojiCell = collectionView.dequeueReusableCell(for: indexPath)
+
+        let url = emojisList[indexPath.row].emojiUrl
+
+        cell.setUpCell(url: url)
+
+        return cell
+    }
+}
+
+// extension EmojisListViewController: EmojiStorageDelegate {
+//    func emojiListUpdated() {
+//        collectionView.reloadData()
+//    }
+// }
