+++ Challange-1-iOS/Challange-1-iOS/Utilities/ImageView+Download.swift
@@ -16,5 +16,4 @@ extension UIImageView {
         }
         return task
     }
-
 }
